Split Github and Bitbucket hook handlers
See #2 #232
